A round of 64-to-32 warning squashes
formatting fixes and type promotions to chip away at the 1500 lines of
warnings that pop up with clang's -Wshorten-64-to-32 flag.

+++ src/mpi/romio/adio/common/ad_fstype.c
@@ -172,7 +172,7 @@ static void ADIO_FileSysType_parentdir(const char *filename, char **dirnamep)
 	 * but this code doesn't care if the target is really there
 	 * or not.
 	 */
-	int namelen;
+	ssize_t namelen;
 	char *linkbuf;
 
 	linkbuf = ADIOI_Malloc(PATH_MAX+1);
